che do lua chon speed, gan nhac va dieu chinh fxml

+++ src/model/SlowFoodEffect.java
@@ -5,7 +5,7 @@ public class SlowFoodEffect implements FoodEffect {
     public void applyEffect(Snake snake, GameConfig gameConfig, GameState gameState) {
         System.out.println("Slow food eaten. Snake slows down.");
 
-        int newSpeed = gameState.getSnakeSpeed() + 100; // Tăng tốc độ
+        double newSpeed = gameState.getSnakeSpeed() + 100; // Tăng tốc độ
         gameState.setSnakeSpeed(newSpeed); // Cập nhật trạng thái
         
     }
